Developing conjunction of clauses
The building is complete returning a bdd list this will then be
injected in to the conjunction module to get a bdd. This will be the
first major mile stone. Then the conjunction bdd will be converted back
in to cnf formula for the second major milestone after which we can run
evaluation.