Segmentation of Conjunction clauses implemented
This method will break up a given expression list given a value x. This
value x is used as an index of breaking up a list and returning this
segment. This segment is then used for conjunction of the clauses in
the list. For example given a list [1;2;3;4;5] an x=2 at index 2 it
will return a new list which is [1;2] and again will return [3;4] and
finally [5] and conjunction modul will be ran 3 times with the
individual lists. A combined list is returned which has implicit
conjunction between them. To test you could run conjunction between
them too.