Still developing new feature
My supervisor convinced me this was 2 hours work. It's taking a lot
longer than anicpated. Deadline in 2 days. This includes a 30 thousand
word thesis. To be fair, I just need the evaluation and finishing
touches in thesis. But still, why ask for this when ther's already so
much pressure! I guess it's for my own good.